requirements.txt
pyyaml==5.4.1
tweepy==4.5.0
pydrive==1.3.1
loguru
psycopg2-binary==2.9.3